Removed most TypeScript related code
