+++ .eslintignore
@@ -1,3 +1,2 @@
 __mocks__
 __snapshots__
-**/styles.ts
