+++ babel.config.js
@@ -25,8 +25,6 @@ module.exports = (api) => {
   return {
     presets: [
       "next/babel",
-      "@zeit/next-typescript/babel",
-      "@babel/preset-typescript",
       "@babel/preset-react",
     ],
     plugins: ["@babel/proposal-class-properties"],
