+++ package.json
@@ -10,7 +10,7 @@
     "ci": "npm run lint && npm run test",
     "deploy": "vercel",
     "dev": "node server.js",
-    "lint": "eslint pages components constants src --ext .ts,.tsx,.js,.jsx",
+    "lint": "eslint pages components constants src --ext .js,.jsx",
     "start": "next start",
     "test": "jest"
   },
@@ -57,8 +57,7 @@
   "devDependencies": {
     "@babel/core": "^7.12.3",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/preset-react": "^7.12.5",
-    "@babel/preset-typescript": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
     "@inrupt/eslint-config-base": "^0.0.4",
     "@inrupt/eslint-config-react": "^0.0.4",
     "@sentry/webpack-plugin": "^1.13.0",
@@ -71,8 +70,8 @@
     "@types/react-table": "^7.0.25",
     "@typescript-eslint/eslint-plugin": "^3.10.1",
     "@typescript-eslint/parser": "^3.10.1",
-    "@zeit/next-typescript": "^1.1.1",
     "babel-eslint": "^10.1.0",
+    "babel-jest": "^26.6.3",
     "eslint": "^7.12.1",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-airbnb-base": "^14.2.0",
@@ -95,7 +94,6 @@
     "prettier": "^2.1.2",
     "raw-loader": "^4.0.2",
     "react-test-renderer": "^17.0.1",
-    "ts-jest": "^26.4.3",
     "typescript": "^4.0.5",
     "vercel": "^20.1.2",
     "whatwg-fetch": "^3.5.0"
