+++ resources/whitesource/wss-pull_request-scan.config
@@ -45,7 +45,7 @@ bower.resolveDependencies=false
 ###########################################################################################
 # Includes/Excludes Glob patterns - Please use only one exclude line and one include line #
 ###########################################################################################
-includes=**/*.js, **/*.ts
+includes=**/*.js
 
 case.sensitive.glob=false
 followSymbolicLinks=true
