+++ resources/whitesource/wss-push-scan.config
@@ -46,7 +46,7 @@ bower.resolveDependencies=false
 # Includes/Excludes Glob patterns - Please use only one exclude line and one include line #
 ###########################################################################################
 
-includes=**/*.js, **/*.ts
+includes=**/*.js
 
 case.sensitive.glob=false
 followSymbolicLinks=true
